DevServer: don't blow up on error reloading module
This wraps the code path which triggers ssrLoadModule in a try/catch.
ssrLoadModule is where any exceptions/errors that occur as a result
of making changes and reloading a page will originate from.

J=SUMO-4390
TEST=manual

Manually tested making breaking changes and reloading
that page. Made sure that I was re-directed to an
error page and that the devserver did not crash.

https://drive.google.com/file/d/1_LOi5661vWZGZkTlbDNlQqjwI-JiudPK/view?usp=sharing