SUMO-4390: Add custom error handling middleware
This adds an error handling middleware [1] to our Express dev server.
This middleware will be invoked anytime a middleware above it
(i.e. all of them) calls the next function with an error. The
HTML page template, styling, and most of the error parsing
logic was taken from this NPM package [2]. Pulling it out
of that package was the only way to do things such that
the Vite HMR client could be injected into the error page.

I put the middleware in the server.ts file because the
favicon middleware was there, but it probably makes sense
to refactor a bit and create a directory for middlewares,
like the favicon one and serverRenderRoute. But, that
can be another PR / commit.

J=SUMO-4390
TEST=manual

https://drive.google.com/file/d/1cIZoB0L8HFp99NQhIDOmNx3rYEOGcnCa/view?usp=sharing

1: https://github.com/senchalabs/connect#error-middleware
2: https://www.npmjs.com/package/errorhandler