Use ansi-to-html Tool
The stack trace looked messed up becuase it was using a bunch of ansi
codes to color / style itself, which the terminal can resolve but
a web browser cannot. This adds the npm ansi-to-html pacakge and
uses it to cleanup the html error page.

The spacing being slighly off is a result of the implementation of
the ansi-to-html and it inserting the spaces between the opening of
two HTML tags instead of inserting them where they'd be rendered.
I tried two other popular ansi to html npm packages and they had
the same issue and looked worse, so for now I think we should just
stick with this.

J=SUMO-4390
TEST=manual

https://drive.google.com/file/d/1odlKzs8zk18_DWmOdUf7XBPtbTw5AmyD/view?usp=sharing