Refactor yext-sites-scripts
This CR tweaks the directory layout for yext-sites-scripts so files
are now grouped by feature. The entry point is still in /bin, but
besides that everything is grouped under the name of the relevant
command. Hopefully this will make development a bit more
straightforward.

J=none
TEST=manual

Ran the various commands and made sure there weren't any errors.