Add custom error handling middleware (#9)
Add custom error handling middleware

This adds an error handling middleware [1] to our Express dev server.
This middleware will be invoked anytime a middleware above it
(i.e. all of them) calls the next function with an error. The
HTML page template, styling, and most of the error parsing
logic was taken from this NPM package [2]. Pulling it out
of that package was the only way to do things such that
the Vite HMR client could be injected into the error page.

I put the middleware in the server.ts file because the
favicon middleware was there, but it probably makes sense
to refactor a bit and create a directory for middlewares,
like the favicon one and serverRenderRoute. But, that
can be another PR / commit.

J=SUMO-4390
TEST=manual

https://drive.google.com/file/d/1cIZoB0L8HFp99NQhIDOmNx3rYEOGcnCa/view?usp=sharing

1: https://github.com/senchalabs/connect#error-middleware
2: https://www.npmjs.com/package/errorhandler

* Address Comments

This commit addresses some minor comments fromt the previous CR.

- group vite imports
- move fixStackTrace to the error middleware

J=SUMO-4390
TEST=manual

Tested the error flow, made sure everything worked as before.

* Use ansi-to-html Tool

The stack trace looked messed up becuase it was using a bunch of ansi
codes to color / style itself, which the terminal can resolve but
a web browser cannot. This adds the npm ansi-to-html pacakge and
uses it to cleanup the html error page.

The spacing being slighly off is a result of the implementation of
the ansi-to-html and it inserting the spaces between the opening of
two HTML tags instead of inserting them where they'd be rendered.
I tried two other popular ansi to html npm packages and they had
the same issue and looked worse, so for now I think we should just
stick with this.

J=SUMO-4390
TEST=manual

https://drive.google.com/file/d/1odlKzs8zk18_DWmOdUf7XBPtbTw5AmyD/view?usp=sharing

* Add comment explaining error handling middleware