Remove Artifacts
The dev, build, and preview commands are called via yss from the
site that's being developed. There doesn't need to be a separate
set of NPM commands in the yext-sites-scripts package for them
anymore. This also allows us to remove the entry-server.ts file
which was almost the exact same as dev.ts.

J=none
TEST=manual

Ran dev and build commands on the starter site against the local
scripts and made sure it worked fine.