Add API-Extractor and API-Documenter
This adds api extractor and documenter in a way that will generate
an api report and markdown docs every time we re-generate types.

The config for api-extractor and api-generator was taken from
Slapshot here [1] and is configured to only generate docs for
methods/types/interfaces that are annotated with @public.

J=SUMO-4391
TEST=manual

Ran npm run types and inspected the output.

1: https://github.com/yext/answers-react-components/blob/main/api-extractor.json