Refactor Sites Scripts
This refactors sites scripts to be organized by feature.
It's the same as the last PR but it was easier to just
make a new one then try to resolve all the conflicts.

J=none
TEST=manual

Manually tested the sites scripts commands with the starter
and made sure they still worked correctly.