Testing: Add jest and an example test
This commit sets up a framework for jest testing in our repo
and adds a test for one of the ssr functions.

In a forthcoming commit I'll add a test that is actually
meaningful, I just wanted to get things configued so that
tests could run.

I also re-ran the types command which updated some of the
auto-genned api documentation.

J=SUMO-4393
TEST=auto

Ran npm run test and made sure that my unit test ran.