VitePlugin: Update build output
This CR updates our Vite plugin to spit out our build output
in a way that greatly simplifies the ci.json file.

I also upgraded rollup because we were on a super ancient version
and it didn't seem to break anything.

To see a site that works with this, here's the max-shaw-site
on the branch I just pushed the accompanying changes to (thought
it would automatically open a PR but it didn't, sorry justin):
https://github.com/yext/max-shaw-sites-react-starter/tree/justins-changes

J=SUMO-4541
TEST=manual

Manually tested the plugin with Max's test site and checked that
it properly generated all the pages and that react worked on those
pages.