Fix Plugin: change path resolution (#45)
The plugin needs the files to be in the actual plugin
folder of the .artifact-output to be able to access them
in prod mode, it won't be able to reach across into the
desktop directory like it is locally. Whoops...

J=none
TEST=manual

Manually tested locally that things still worked. Need
to cut new releases to test in yext sites.