Support the global variable (#49)
* Support the global variable

This aliases the global variable with globalThis. A problem
was being called where some third party node packages were
relying on the presence of global, but in deno it is not
provided and instead globalThis is used.

J=none
TEST=manual

Replicated the error by console.log-ing global, made sure
that it didn't cause an error anymore after adding this
rollup config option. Also checked the bundled output
to make sure it was properly referencing the added piece
of code.

* [format-bot] apply prettier formatting changes

Co-authored-by: jimboHenris <jimboHenris@users.noreply.github.com>