SitesScripts: Support arbitrary user defined head content
This adds a type called HeadConfig and allows a user to define
it in their template in such a way that allows them to entirely
arbitrarily set the inner content of the head element.

J=SUMO-4579
TEST=auto,manual

Added and ran unit tests.

Manually tested defining all parts of the HeadConfig and made
sure that the generated HTML documents in dev and prod mode
were correct.