Documentation: Fix
This fixes some issues with the autogenned documentation.
In order to fix one issue a config setting for vitepress
needed to be set which was only included in the newest release,
so this bumps up to v1.0.0-alpha of vitepress.

TEST=manual

npm run docs:build && npm run docs:serve

Made sure the doc site and all the autogenned pages
worked properly.