YSS: Create an index page for the dev server
This updates the dev server to generate an index.html page which
will be automatically opened when the server is started. The page
will contain an info message depending on if the user is running
in local or dynamic mode and will display a selection of URLs
based on the content of the user's localData folder for their
site.

If run in dynamic mode, the test data will automatically be
refreshed on dev server start. There will also be a warning
message if this refresh fails for some reason.

Behold, a truly responsive design [1]

1: https://drive.google.com/file/d/1s2AzlV6jBNsrlyRG18cQg-gxPoWsnu0f/view?usp=sharing

TEST=manual
J=SUMO-4412

Tested the page with the starter and made sure that it worked
for static as well as entity pages.

Tested the various failure states (starting without a localData
directory in local and dynamic mode, starting without a network
connection in dynamic mode, etc.) and made sure that the expected
warnings / errors were displayed in each case.