SitesScripts: Add a relativePrefixToRoot
This PR creates a new type that we'll pass into the template when
rendering called TemplateRenderProps that extends the TemplateProps
interface with the additions of a path and a relativePrefixToRoot.

TEST=auto,manual
J=SUMO-4617

Added and ran unit tests.

Manually tested in dev and prod mode. Made sure for paths of
varying lengths that the relativePrefixToRoot was correct.