YSS: Start dev server in dynamic mode by default
This is currently broken with static pages, but Duval has a fix
coming in his PR so I'll just wait to merge until that's merged.

Also had to update the generateTestData command to work with the
new CLI.

J=SUMO-4620
TEST=manual

Started dev mode, made sure it was in dynamic. Also, made sure
the -- local flag worked as expected.