Fix ts compiler error
When trying to use ProjectStructure in a test, npm run test was
producing a ts error that was due to us overloading the name
"path" in the getRelativePath and getAbsolutePath functions.

TEST=auto

declared a ProjectStructure in generateTestData.test.ts and
made sure that this let the tests run without throwing the
ts error.