YSS: Guard head tag
Print a warning message to console when a tag without a type is being rendered.
Note: other places seem to be guarded already so this seems to be the only addition I can make to fulfill this item.

J=SUMO-4625
Test=manual

Tested by adding a tag without a type to getHeadConfig and make sure that warning messages are printed.