YSS: Guard head tag  (#92)
* added test and changed from chalk package to picocolors

* YSS: Guard head tag

Print a warning message to console when a tag without a type is being rendered.
Note: other places seem to be guarded already so this seems to be the only addition I can make to fulfill this item.

J=SUMO-4625
Test=manual

Tested by adding a tag without a type to getHeadConfig and make sure that warning messages are printed.

* YSS: Guard head tag

Print a warning message to console when a tag without a type is being rendered.
Note: other places seem to be guarded already so this seems to be the only addition I can make to fulfill this item.

J=SUMO-4625
Test=manual

Tested by adding a tag without a type to getHeadConfig and make sure that warning messages are printed.

* YSS: Guard head tag

Changed warning message to address tag type is unsupported

J=SUMO-4625

* WIP: guard head tag

* changed warning message

* fix warning message

* added a test rendering tag with upsupported type, changed chalk package to picocolors

* added a test rendering tag with upsupported type, changed chalk package to picocolors

* addressed Duval's comment -- mainly added a test that console logs the warning when tag type is not supported

* ran npm run release

* added an empty line to separate the tests in head.test.js

* [format-bot] apply prettier formatting changes

* Automated update to THIRD-PARTY-NOTICES from github action's 3rd party notices check

Co-authored-by: anguyen-yext <anguyen-yext@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>