fix: update getRuntime to work with the plugins system
We noticed a bug where our getRuntime function makes the
assumption that a global window object will be available.
This will not be true when running in the plugins system
and will cause page generation to fail.

TEST=auto,manual

Updated unit tests. Tested locally to make sure getRuntime
still returns the expected results for node, deno, and
browser.