fix(dev/plugin): address incompatibilities with windows (#173)
This CR handles a number of incompatibilities the PagesJS toolchain had with windows and updates it to make it functional. Specifically, I've updated entry-points in our code which touch the underlying operating system to be path separator aware using `path.join`. 

I've made the opinionated decision that we only want these entry-points to change while keeping the internals of the system using posix-style separators. This is because node requires posix-style separators for imports and mixing them internally seems prone to cause more unintentional bugs in the future. This notably means that even on windows, the paths generated by our `manifest.json` still use posix-style separators. Considering we use the manifest internally to import and match node imports this consistency is important.

Fixes: https://github.com/yext/pages/issues/160
J=SUMO-4729
TEST=auto,manual
 
Ran tests and tested the following on windows and unix:
- hydration 
- prod mode page generation
- dev mode page generation (dynamic)
- dev mode page generation (local)
- stateful updating

Co-authored-by: Duval Kilpatrick <mkilpatrick@yext.com>
Co-authored-by: mancusi <mancusi@users.noreply.github.com>