fix(components): make fixed Image width and height optional (#241)
Fixes #195.

J=SLAP-2400
TEST=auto, manual

See that the updated Jest test passes and tested locally to make sure a `width` could be specified without a `height` and vice versa when the image had a fixed layout. Specifying neither `width` nor `height` or invalid values for either still logged an error, as did specifying either a `width` or `height` (or both) when the image layout was not fixed.