feat(components): add `LocationMap` component (#242)
This PR adds the `LocationMap` component, which allows uses to easily
add a single marker to the map and have it link to another page on
marker click. The common use case for this is adding it to a location
page and having the pin click link out to the directions.