feat(dev): document.slug URLs in dev mode (#256)
This PR changes URLs in dev mode from `/[template-name]/[entity-id]` to the form `/[document.slug]`,
where `document.slug` is the slug field in the entity and corresponding test data.

In order to render a template, we need two things, the template file itself (for example `src/templates/location.tsx`),
and the data for the entity we want to render using the template, i.e. the test data.

With the original URLs, the template to use is just the [template-name] in the URL, and the test-data is fetched
via the Yext CLI using the [entity-id] in the URL combined with the `config` export that could be loaded from the template.
The `config` would then be passed to the CLI inline.

With the new URLs, the order we get this information has flipped around. With prod URLs, we first get the test data
by hitting the CLI using the new --slug parameter added to generate-test-data in https://gerrit.yext.com/c/alpha/+/192814.
However, we STILL need a FeaturesConfig to send to the CLI. At least I think we do, when I try to call the CLI without it I get an error.

Since we have no way of knowing which template we need yet, and can't just get the `config` export like the original URLs, I decided to loop through every template and generate a FeaturesConfig with all templates that are for entity pages. I'm not sure if this is the right way to do this but it seems to work? This seems to let us hit the CLI properly, and gets us the test data we want. The CLI should error out if it encounters 0 or multiple entities for a single slug. 

(I believe this approach will not work if multiple templates share an entity, but I don't think that's a case we can address without more information, because which template would we even render to the page?)

After getting the test data we can get the corresponding template using `document.__.name`, and follow the same flow as the original URLs.

Also fixes a bug where static pages in --local mode would sometimes grab the wrong test data document (they would always grab the first one that did not have an entityId)

Adds an error message to the 404 pages for cases like when no document with the slug is found.

J=SLAP-2447
TEST=manual

when --no-prod-url is set, the original URL scheme will still be used. tested with --local and also dynamically generated test data. index page URLs look correct for both static and entity pages

tested that with prod urls and --local set the page renders as expected, and the index page has the right links.
tested on https://github.com/yext/pages-starter-react-locations, and also added an additional page in order to test with multiple entity types
tested that alternativeLanguageFields still works and will provide a document for the right locale with alternativeLanguageFields 
tested that index page will render (relatively) gracefully even when localData does not have slugs