dev(fix): add back favicon support in dev mode
It appears we can't use loadTemplateModules() in dev mode.
If we try to do so we lose support for importing .ico files
(and I imagine other kinds of loaders too).

Instead, calling loadTemplateModule() (which is not used in loadTemplateModules())
in a loop works since it relies on vite to import modules.

J=None
TEST=manual

tested that I can import .ico favicons in the pages react starter,
both dev mode and prod mode (yext pages generate -> yext pages serve) work