chore: update GH workflows to remove deprecation warnings (#262)
Upgrade GitHub actions (`pnpm/action-setup`, `amannn/action-semantic-pull-request`, and `EndBug/add-and-commit`) to remove all deprecation warnings from the workflows. These warnings were for the deprecation of [Node 12](https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/) and for the `set-output` and `save-state` [commands](https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/).

J=SLAP-2427
TEST=manual

Test in this PR as well as #260 and #261. See that the warnings are no longer present. Note: the linting workflow will still show warnings until this PR is merged in.