chore(dev/plugin): add file extension to node_module/file import
file extensions are required for relative imports but not for node_modules. react-dom/server is unique because it is a node_modules import, but it references a specific file. This is fine in Node 16-18, but errors in Node 19. This item fixes support for Node 19

J=SLAP-2501
TEST=manual

successfully served the site using Node 16, 17, 18, and 19