fix(dev): use Vite JS API for build command (#288)
Use the Vite JS API instead of its CLI for the `pages build` command. This ensures that the version of Vite that's used to create the build is one that matches the PagesJS `vite` dependency, even if another version of Vite is installed in the repo.

J=SLAP-2621
TEST=manual

Use a packed version of PagesJS in a version of the locations starter running Vite 4 and see that Vite 3 is still used when `npx pages build` is run. Also tested that the scope flag still works for a multibrand repo.