feat(plugin): generate functionMetadata.json (#293)
This PR adds a step to the vite-plugin's `build` command (i.e. `pages build`) to generate `functionMetadata.json`. This file contains a mapping of the files in the `functions` directory to the names of the functions that are their default exports. If there is no top-level `functions` directory, `functionMetadata.json` is not generated.

J=SLAP-2625
TEST=manual

Point the locations starter to a local pack of PagesJS and see that the file is correctly generated when there is a `functions` directory, and is not generated if there is no `functions` directory.