test(playground): upload Windows/Linux screenshots via CI (#302)
In order to run snapshot testing in Windows and Linux, snapshots for those environments must be committed to the repo.
Since we don't want to manually add snapshots from 3 different operating systems, this PR updates the playwright ci action to do that for us.

The action now only runs on push instead of on pull_request. When the playwright action runs on pull_request, we're not able to push to the original branch. This is because actions that run on pull_request are run on a new branch created automatically by github, this new branch is the merge commit between the head and base branch.
Running on push will run the action directly on the head branch.

J=none
TEST=manual

see that snapshots are uploaded from ubuntu and windows