feat(dev/plugin)!: allow custom client/server rendering templates (#309)
This adds the ability for a user to define their own
_client.tsx/_server.tsx so that they have full control over the server
and hydration rendering. If none are defined in the user's repo, it will
default to built-in templates provided by PagesJS.

This is a **breaking change** for anyone not pinned to a specific
previous version. This is because a new folder _render_ is output during
the build process that YextCI must be aware of. The Generator plugin
section in ci.json should now look like this:
```
{
  "pluginName": "Generator",
  "sourceFiles": [
    {
      "root": "dist/plugin",
      "pattern": "*{.ts,.json}"
    },
    {
      "root": "dist",
      "pattern": "assets/{server,static,renderer,render}/**/*{.js,.css}"
    }
  ],
  "event": "ON_PAGE_GENERATE",
  "functionName": "Generate"
}
```
_Note that a future PR will auto-generate the Generator ci.json section
during build, making this a moot issue._

This change also has the added benefit that only a single hydration
template (_client.tsx) is necessary. Previously, every template built
required its own hydration template. This should cut down on the number
of files necessary by the Yext Plugin system as it is now rendered at
runtime.