feat(dev/plugin): add Generator plugin to ci.json during build (#311)
There is a new pages generate ci command that will auto update the
ci.json's Generator plugin. It also applies during the build step. This
is because PagesJS understands how it builds its output to work on the
Yext platform and shouldn't require a user to understand the details.
Now they don't need to.