feat(plugin): fail the build when there are errors (#314)
Before this change a step in the build process could fail but because it
didn't throw the YextCI build system would view it as successful,
leading to confusion. If any step has an error we now throw so it's
clear what happened and won't make its way to the Generation phase.