feat(dev/plugin)!: added hydration option to TemplateConfig (#327)
Right now, hydration is set to true by default. This PR adds a hydration
field to `TemplateConfig` field (to both internal and externals) and
makes it treated as false by default. It is an optional argument, so
undefined is treated as false (default).

This was tested by setting hydrate to `true` and `false` (both dev and
with `build:serve` prod) and a react component that counted how many
clicks was placed. When set to true, the counter worked and when set to
false, the counter did not (as expected).