fix(dev): add additional filetypes to esbuild loader for features.json (#331)
The features.json generation was failing in some circumstances when
unknown filetypes were being imported into the templates. This is only
an issue when running `pages features` directly and does not affect the
build process. It's possible we may want to just remove this feature
altogether as its own command.