fix(dev): index page QA - update warning/info text and add favicon (#333)
- adds Yext favicon to the dev index page
- updates the info text in the sidebar
- removes the generate test data failed error/bug
- adjust warning div styles
- allows getPath to have non-standard URL characters, particularly a
space. Limitation: a getPath with `%` in it will crash.

### Encoded URI Strings
Vite seems to have inconsistent uses of URI encoded and decoded strings.
`vite.transformIndexHtml` requires a valid URI string because it runs
decode to get the html file path. However, it caches the transformations
as the literal string provided and then looks them up using a decoded
string.

**For example:**
Previously if `getPath` returned `'a b c'` it would be passed to
`vite.transformIndexHtml` as `'a%20b%20c%20'`. The html file would be
looked up as `path/to/abc.html` but the transformation would be stored
as `'a%20b%20c%20'`. Then, the html transformation would look up `'a b
c'`, creating the the bug. Now, the path is decoded before being sent to
`vite.transformIndexHtml`, so it receives as `'abc'`, which passes
through the decoder fine for the html and then also works for the
transformation lookup.

The new limitation is that if `getPath` returns `'abc%'` the decoder
throws an error because that is not a valid URI string. Similarly, if
`getPath` returns an encoded string like `'abc%20'`, the `%` it gets
encoded by the server creating `'abc%2520'` which then causes the
original bug to occur.