refactor(components)!: move components out to go into @yext/pages-components (#361)
Deleted everything in pages/src/components because it was moved to a
separate repo and will be packaged as @yext/pages-components. Saw that
there were no dependencies within the pages repo for components. Tested
by packing a local version of this, and packing a local version of the
new @yext/pages-components package, importing them into a local test
repo, and seeing that components rendered properly (except for
<LocationMap> but that never renders properly when local packing
package).

Removed dependencies in `package.json` and any comments / README / other
stuff referencing components.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>