fix(dev/plugin)!: redo reverse proxy setup (#376)
Prior to this change, assets were imported relatively, also taking into
account relativePathToPrefix. While this generally worked, there were
other issues due to Vite's build optimizations: specifically module
preloading and inline assets. Due to this we decided to take a new
approach where assets are now imported absolutely and it's now up to the
RP config to handle the assets. A working example of a Netlify setup for
a RP at the `subdirectory` subfolder looks like this:

```
[[redirects]]
  from = "/subdirectory/assets/*"
  to = "https://rp-test.pagesprod.yextengtest.com/subdirectory/assets/:splat"
  status = 200
  force = true
  headers = {Host = "https://rp-test.pagesprod.yextengtest.com/"}
  
[[redirects]]
  from = "/subdirectory/*"
  to = "https://rp-test.pagesprod.yextengtest.com/:splat"
  status = 200
  force = true
  headers = {Host = "https://rp-test.pagesprod.yextengtest.com/"}
  
[[redirects]]
  from = "/subdirectory"
  to = "https://rp-test.pagesprod.yextengtest.com/index.html"
  status = 200
  force = true
  headers = {Host = "https://rp-test.pagesprod.yextengtest.com/"}
```

In order to properly configure the Vite side, a user must set the
assetsDir in the vite.config.js (`subdirectory/assets` for the above
example) or in the config.yaml (which will be globally supported in some
upcoming PRs).

Other changes in the PR include a massive refactor of ProjectStructure
in order to simplify its usage and is now being used everywhere, meaning
that all filenames and folders are fully configurable.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>