fix: serverless function imports (#384)
This changes how serverless functions are compiled. Previously, they
were being bundled through Vite via Rollup. This did not allow proper
bundling because both the templates and functions resulted in shared
assets. It also presented an issue with relative imports but the final
functions needing to live in a non-relative place to those assets in
.artifact-output.

Now, serverless functions are handled during closeBundled, with each
function being bundled independently from each other (and templates). We
use esbuild to do the bundling.