fix(dev): use ts-morph to parse template configs (#387)
Previous to this change, we used esbuild to compile the templates into
modules in order to pull off the config const. This led to various
issues for people. It works fine with Vite since Vite does a lot of
CJS/polyfill magic, but we weren't handling that with esbuild. This was
overkill anyways as we only need to grab the config off of the
templates, so compilation is irrelevant at this step.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>