revert(dev): use esbuild instead of ts-morph to parse template configs (#396)
This is a manual revert to use esbuild again instead of ts-morph because
the latter doesn't compile anything. This led to issues when a template
used a function inside of its exported config.

The change for esbuild now is that there is a fallback loader to load
any unknown filetype using the `file` esbuild Loader.