fix(plugin): remove import.meta.resolve dependency
Removes the dependency on import.meta.resolve which
is only supported in the newest versions of Safari.

It ultimately was unnecessary as pagesJS was providing
absolute paths to the call which returns a URL to the module.
These get resolved during an import in the same
manner as an absolute path so it can just removed.

Before:
```
// url = "https://domain.com/abs/path"
const url = import.meta.resolve("/abs/path")
// import("https://domain.com/abs/path")
import(url)
```
After
```
import("/abs/path")
```

TEST=auto,manual

ran unit tests. Manually tested on a RP site and a non-RP site.
Since the asset paths are absolute, behavior is the same
assuming the RP is properly forwarding assets.