fix(dev): use Vite instead of esbuild, also Vitest (#416)
This is _yet another_ change in how we generate features/templates. The
story have been to use esbuild, then ts-morph, back to esbuild and now
Vite. We use the Vite devServer directly to handle compiling the code so
we can import the template module to get the config off of it.

This change also replaces Jest with Vitest, which allows the testing of
esmodules much more easily. It was also required in order to actually
run Vite in some tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>