fix(dev/plugin): support HTML in Content fields (#422)
Fixes #225.

To support this, we now encode the page props before injecting the JS bundle in the DOM and have the JS bundle decode the props before passing it to the `render` function that renders the page.

J=SUMO-5337
TEST=manual

In a test site, see that both `dev` and `build:serve` can successfully serve a page that is using a Content field with HTML in it that includes a `</script>` closing tag. Test in the platform as well and see the page has errors without this change, and renders properly with the fix.