fix(dev): include document support for static page `getPath` functions. (#421)
Fixes https://github.com/yext/pages/issues/415. 

In the past, if users tried to access `document` fields in a static page
`getPath` function, it would show up as "undefined" in the page URL.
This PR addresses this by giving static pages access to the `document`
object.

A couple changes were made to accommodate this functionality. We want to
pass in a document in order correctly compute a `getPath` function. This
means that when we iterate through templates to find the corresponding
template document to our URL, (see `findMatchingStaticTemplate`), we
need to retrieve the documents which is an asynchronous action.
Therefore, the `criterion` function in `findTemplateModuleInternal` was
changed to accept async functions.

For clarity, `urlToFeature` was also refactored. It was also found that
`defaultLocale` was not being used in `serverRenderRoute`. Before, we
hard-coded `en` to be the default Locale in `urlToFeature`. But in this
refactor, we use `defaultLocale` instead.

J=SUMO-5417
TEST=manual