fix(dev/plugin): clearer error for non-string getPath value (#425)
Fixes #423.

Now, we explicitly check for and throw an error with a more clear message if the `getPath` value is not a string.

J=SUMO-5458
TEST=manual

In a test site in dev mode, see that the index page does not include any pages with a non-string `getPath` value and console error is logged instead. In `--no-prod-url` mode, see that visiting an entity page with a non-string `getPath` value throws an error with the more specific message instead of `path2.split is not a function`. Running `yext pages build && yext pages render` locally, as well as page generation in the platform, will throw the clearer error for a non-string `getPath` rather than `n.split is not a function`.