feat(dev/plugin): add localization to static pages (#430)
This PR adds localization to static pages and adds an interface for
`StaticTemplateConfig`. Localization can now be specified for static
pages by including the following in the `TemplateConfig`.

```
export const config: StaticTemplateConfig = {
  name: "robots",
  locales: ["en", "es"]
};
```

This config is included in the `templates.json`/`features.json` to be
consumed by Spruce as:

```
{
  "features": [
    {
      "name": "robots",
      "templateType": "JS",
      "staticPage": {
        "locales": ["en", "es"]
      }
    },
  ],
  "streams": []
}
```

Additionally, it was found that since static pages no longer have a
single `staticURL` per `featureName`, the `localDataManifest` had to be
refactored. This is because static pages now accept a `document` in
their `getPath` function, meaning that when static pages may have
different URLs depending on the document if the path uses the
`document`'s `locale`.

Furthermore, it was found that when running `yext pages render` to
generate pages, static pages with a constant `getPath` return value
multiple locales would only generate one static page. To deal with this
behavior, pages will now throw an error when users have multiple static
pages with identical paths.

J=SUMO-5451
TEST=manual

It was tested that you can specify `locale`s in the `TemplateConfig` for
static pages, and that they property show up in the `templates.json`. It
was also tested that when `yext pages generate-test-data` is called, it
generates a document for each locale specified in a static page.

---------

Co-authored-by: Alexis Sanehisa <106991746+asanehisa@users.noreply.github.com>