docs: auto generate markdown files  (#441)
This commit adds a new workflow, generate markdown files, which runs
api-extractor and api-documenter in order to generate markdown files for
all the @public functions in our src code.

There is now a bot that auto runs tests after the generate markdown file
workflow and check + update 3rd party notices workflow.

How the bot works: 
The bot has write access to the repo, so it is like a regular user
committed to the repo, rather than an automated commit. GH prevents
automated commits from a workflow, which uses the GITHUB_TOKEN by
default, from re-running checks (b/c it might cause an infinite loop of
GH actions running), but commits from a workflow with a personal access
token will trigger a re-run of tests. so, in a way we're bypassing some
of the built-in GH protections, so we have to be careful to only use the
bot token in a repo when we know for sure that it won't create an
infinite loop of GH workflow triggering.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>